fixed-sqpam-bug

+++ src/utils.py
@@ -22,7 +22,7 @@ def convert_to_probability_amplitudes(array):
 	return norm, probability_amplitudes
 
 def convert_to_angles(array):
-	return np.arcsin(np.sqrt((array+1)/2))
+	return np.arcsin(np.sqrt((array.astype(float)+1)/2))
 
 def get_time_resolution(array):
 	time_resolution = int(np.ceil(np.log2(len(array))))
